examples/producer/requirements.txt
kafka-python==2.0.2
faker==22.5.1
pycountry==23.12.11
